+++ package.json
@@ -25,6 +25,7 @@
     "mongoose": "^5.10.9",
     "passport": "^0.4.1",
     "passport-facebook": "^3.0.0",
+    "passport-google-oauth20": "^2.0.0",
     "passport-local": "^1.0.0"
   }
 }
