+++ public/stylesheets/style.css
@@ -1,1 +1,21 @@
+body {
+  background: #f8f8fd;
+  color: #514B64;
+  /* min-height: 100vh; */
+}
 
+.btn-circle {
+    position: relative;
+    width: 35px;
+    height: 35px;
+    border-radius: 17px;
+    text-align: center;
+    font-size: 12px;
+}
+
+.fab {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
