+++ views/login.hbs
@@ -13,10 +13,13 @@
           <label for="password">Password</label>
           <input type="password" id="password" name="password" class="form-control" placeholder="Enter password...">
         </div>
-        <button type="submit" class="btn btn-primary btn-block">Login</button>
-        <a href="/auth/facebook" class="btn btn-info btn-block">FB Login</a>
+        <button type="submit" class="btn btn-info btn-block">Login</button>
+        <div class="d-flex justify-content-around mt-2">
+          <a href="/auth/facebook" class="btn btn-primary btn-circle"><i class="fab fa-facebook fa-2x"></i></a>
+          <a href="/auth/google" class="btn btn-danger btn-circle"><i class="fab fa-google fa-2x"></i></a>
+        </div>
       </form>
-      <p class="lead mt-3">
+      <p class="lead mt-1">
         <small>Don't have an account? <a href="/users/register">Register</a></small>
       </p>
     </div>
